In this revision,    SatID is fully implemented in the library and applications,    all the calls to functions in the various EphemerisStore classes       are upgraded to the SatID version,    RinexObsHeader::RinexPrn is removed entirely, (although enum RinexSystem       remains in RinexObsHeader), and instances of RinexPrn are changed to SatID,    RinexObsHeader::RinexPrnMap    are renamed   RinexObsHeader::RinexSatMap,    RinexObsHeader::numObsForPrn                 RinexObsHeader::numObsForSat,    RinexObsHeader::prnList                      RinexObsHeader::satList,    and    RinexObsHeader::lastPRN is changed from int to SatID
Thus enum RinexObsHeader::RinexPrn::RinexSystem becomes RinexObsHeader::RinexSystem,
and 'systemGPS', 'systemGlonass', etc. appear 3 different places
but are always qualified:
   SatID::systemGPS
   SP3Header::systemGPS
   RinexObsHeader::systemGPS
Note that the list of systems in each of these classes is different:
SatID has all the satellite systems but not the file system 'systemMixed'
and SP3 supports more and different satellite systems than Rinex does.
The SP3 and Rinex classes throw when given an unsupported system.

To facilitate using SatID in applications, two additional classes which inherit
from SatID are defined:
   class RinexSatID
   class SP3SatID
These differ from SatID only in that extra functionality is implemented, namely
input from a string (fromString()), constructor from a string, systemChar() output,
and the fill character used in output is accessible and can be changed.
Note that these classes are not used within the toolkit library.

Changes to apps/bindings and apps/reszilla in support of this change
will be made in another revision.


git-svn-id: https://gpstk.svn.sourceforge.net/svnroot/gpstk/dev@107 108fab2b-820f-0410-9f2c-e5cc2c41d7be